# docs/swagger.yaml
openapi: "3.0.3"
info:
  title: Auction Platform API
  description: |
    Real-time auction backend — REST + WebSocket microservices.

    **Auth:** JWT Bearer token in `Authorization` header.
    WebSocket endpoints also accept `?access_token=<token>` query parameter.
  version: "1.0.0"
  contact:
    name: habibiazmi123
    url: https://github.com/habibiazmi123/auction
servers:
  - url: http://localhost:8080
    description: API Gateway (local dev)
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Problem:
      type: object
      properties:
        code:
          type: string
          example: invalid_input
        message:
          type: string
          example: request body is not valid JSON
        request_id:
          type: string

    # ── Auth ──
...
    ProductCreate:
      type: object
      required: [name, description, quantity]
      properties:
        name:
          type: string
          minLength: 1
        description:
          type: string
          minLength: 1
        quantity:
          type: integer
          minimum: 1
...
    Product:
      type: object
      properties:
        id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        name:
          type: string
        description:
          type: string
        quantity:
          type: integer
        status:
          type: string
          enum: [available, locked]
        auction_id:
          type: string
          format: uuid
          nullable: true
        version:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    ProductList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Product"
        page:
          type: integer
        page_size:
          type: integer
        total:
          type: integer
          format: int64

    # ── Auction ──
    AuctionCreate:
      type: object
      required:
        - product_id
        - starting_price_cents
        - minimum_increment_cents
        - starts_at
        - ends_at
      properties:
        product_id:
          type: string
          format: uuid
        starting_price_cents:
          type: integer
          format: int64
          minimum: 0
        minimum_increment_cents:
          type: integer
          format: int64
          minimum: 1
        starts_at:
          type: string
          format: date-time
          description: Must be before ends_at
        ends_at:
          type: string
          format: date-time
          description: Must be after starts_at
        anti_sniping_window_seconds:
          type: integer
          format: int64
          minimum: 0
          default: 0
        anti_sniping_extension_seconds:
          type: integer
          format: int64
          minimum: 0
          default: 0
    Auction:
      type: object
      properties:
        id:
          type: string
          format: uuid
        product_id:
          type: string
          format: uuid
        seller_id:
          type: string
          format: uuid
        product_name:
          type: string
        product_description:
          type: string
        product_quantity:
          type: integer
        status:
          type: string
          enum: [draft, scheduled, live, closed]
        starting_price_cents:
          type: integer
          format: int64
        current_price_cents:
          type: integer
          format: int64
        current_winner_id:
          type: string
          format: uuid
          nullable: true
        minimum_increment_cents:
          type: integer
          format: int64
        starts_at:
          type: string
          format: date-time
        ends_at:
          type: string
          format: date-time
        anti_sniping_window_seconds:
          type: integer
          format: int64
        anti_sniping_extension_seconds:
          type: integer
          format: int64
        version:
          type: integer
          format: int64
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time

    # ── Bid ──
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        type:
          type: string
        auction_id:
          type: string
          format: uuid
          nullable: true
        title:
          type: string
        body:
          type: string
        read:
          type: boolean
        created_at:
          type: string
          format: date-time
    NotificationList:
      type: object
      properties:
        notifications:
          type: array
          items:
            $ref: "#/components/schemas/Notification"

    # ── WebSocket Event Envelope ──
...
  responses:
    400:
      description: Invalid input
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    401:
      description: Unauthorized — missing or invalid token
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    403:
      description: Forbidden — insufficient role or not owner
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
...
    409:
      description: Conflict
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    500:
      description: Internal server error
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
    503:
      description: Service / dependency unavailable
      content:
        application/problem+json:
          schema:
            $ref: "#/components/schemas/Problem"
security:
  - BearerAuth: []
paths:
...
  /v1/products:
    post:
      tags: [Products]
      summary: Create a product (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProductCreate"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "500":
          $ref: "#/components/responses/500"

    get:
      tags: [Products]
      summary: List products (paginated)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            minimum: 1
            default: 1
        - name: page_size
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        "200":
          description: Paginated product list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProductList"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"
...
  /v1/auctions:
    post:
      tags: [Auctions]
      summary: Create an auction (seller only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuctionCreate"
      responses:
        "201":
          description: "Auction created (status: scheduled)"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Auction"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "403":
          $ref: "#/components/responses/403"
        "409":
          $ref: "#/components/responses/409"
        "500":
          $ref: "#/components/responses/500"
        "503":
          $ref: "#/components/responses/503"
...
  /v1/notifications:
    get:
      tags: [Notifications]
      summary: List current user's notifications
      description: Returns notifications for the authenticated user (user ID from JWT).
      responses:
        "200":
          description: Notification list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotificationList"
        "400":
          $ref: "#/components/responses/400"
        "401":
          $ref: "#/components/responses/401"
        "500":
          $ref: "#/components/responses/500"

  # ═══════════════════════════════════════════════
  # WEBSOCKET
  # ═══════════════════════════════════════════════
...